+++ hurocon/cli/auth.py
@@ -1,3 +1,4 @@
+import click
 from click_didyoumean import DYMGroup
 
 from ..implementation import iauth
@@ -11,9 +12,18 @@ def auth():
 
 
 @auth.command('login')
-def auth_login():
-    """ Safely configure all authentication related details for further interactions """
-    iauth.auth_login_impl()
+@click.argument('username', required=False, type=str)
+@click.argument('password', required=False, type=str)
+@click.argument('connection_address', required=False, type=str)
+def auth_login(username: str, password: str, connection_address: str):
+    """
+    Safely configure all authentication related details for further interactions
+
+    To provide all details with one command, specify the USERNAME and PASSWORD
+    positional arguments. CONNECTION_ADDRESS is optional, and will be set to
+    default value (192.168.8.1) if not provided.
+    """
+    iauth.auth_login_impl(username, password, connection_address)
 
 
 @auth.command('logout')